refactor(security-token): :fire: remove nonce from signed transaction
