+++ package.json
@@ -40,6 +40,7 @@
     "@nestjs/platform-express": "^10.3.4",
     "@nestjs/typeorm": "^10.0.2",
     "@noble/secp256k1": "^1.7.1",
+    "@openzeppelin/contracts": "^5.2.0",
     "@orbs-network/ton-access": "^2.3.3",
     "@polkadot/api": "^11.0.2",
     "@slack/webhook": "^7.0.4",
@@ -74,6 +75,7 @@
     "redlock": "^5.0.0-beta.2",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
+    "solc": "^0.8.20",
     "tiny-secp256k1": "^2.2.3",
     "ton": "^13.9.0",
     "tronweb": "^6.0.0",
