+++ src/contract-signer/contract-signer.factory.ts
@@ -1,13 +1,14 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { NonceManagerService } from 'src/nonce-manager/nonce-manager.service';
 import { ContractSignerStrategy } from './strategies/abstract-contract-signer.strategy';
 import { getChainFromNetwork } from 'rox-custody_common-modules/blockchain/global-commons/get-network-chain';
 import { NetworkCategory } from 'rox-custody_common-modules/blockchain/global-commons/networks-gategory';
-import { EVMContractSignerStrategy } from './strategies/evm-contract-signer.strategy';
+import { EVMContractSignerStrategy } from './strategies/evm-contract-signer/evm-contract-signer.strategy';
 
 @Injectable()
 export class ContractSignerFactory {
-  constructor(private readonly nonceManager: NonceManagerService) {}
+  constructor(
+    private readonly evmContractSignerStrategy: EVMContractSignerStrategy,
+  ) {}
 
   async getContractSignerStrategy(
     networkId: number,
@@ -17,7 +18,7 @@ export class ContractSignerFactory {
     let strategy: ContractSignerStrategy;
     switch (network.category) {
       case NetworkCategory.EVM:
-        strategy = new EVMContractSignerStrategy(this.nonceManager);
+        strategy = this.evmContractSignerStrategy;
         break;
       default:
         throw new BadRequestException('Token type is not supported');
