+++ src/signing-transaction/signing-transaction.module.ts
@@ -5,14 +5,16 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { PrivateKeyNonce } from '../nonce-manager/entities/nonce.entity';
 import { BlockchainModule } from 'src/blockchain/blockchain.module';
 import { SigningTransactionRmqController } from './signing-transaction.controller.rmq';
+import { ContractSignerModule } from 'src/contract-signer/contract-signer.module';
 
 @Module({
   imports: [
     KeysManagerModule,
     TypeOrmModule.forFeature([PrivateKeyNonce]),
-    BlockchainModule
+    BlockchainModule,
+    ContractSignerModule,
   ],
   controllers: [SigningTransactionRmqController],
-  providers: [SigningTransactionService]
+  providers: [SigningTransactionService],
 })
-export class SigningTransactionModule { }
+export class SigningTransactionModule {}
